Move the remaining toolchain classes into its own package
This moves the provider and the constants classes into the newly
created toolchain package.
